+++ lib/src/animation/platform_controllers/ios_keyboard_animation_controller.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/animation.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/physics.dart';
-import 'package:keyboard_attachable/src/controller/keyboard_attachable_controller.dart';
+import 'package:keyboard_attachable/src/animation/keyboard_animation_controller.dart';
 
-class IOSAttachableController implements KeyboardAttachableController {
-  IOSAttachableController({@required TickerProvider vsync})
+class IOSKeyboardAnimationController implements KeyboardAnimationController {
+  IOSKeyboardAnimationController({@required TickerProvider vsync})
       : _spring = const SpringDescription(mass: 8, stiffness: 1, damping: 4.5),
         _springVelocity = 10 {
     _controller = AnimationController(vsync: vsync);