+++ lib/src/visibility/keyboard_visibility_controller.dart
@@ -1,0 +1,3 @@
+abstract class KeyboardVisibilityController {
+  Stream<bool> get onChange;
+}